# dataset_servers-main/dataset_servers-main/POS-SYSTEM/openapi.yml
openapi: 3.0.0
info:
  title: Point of Sale (POS) System API
  description: A mock API for a point of sale system with authentication, inventory management, sales processing, customer management, and payment processing.
  version: 1.0.0
servers:
  - url: http://localhost:7000
    description: Local development server
components:
  securitySchemes:
    apiAuth:
      type: apiKey
      in: header
      name: x-auth-token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cashier]
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
...
    Sale:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
                format: uuid
              name:
                type: string
              quantity:
                type: integer
                minimum: 1
              price:
                type: number
                format: float
              subtotal:
                type: number
                format: float
        cashierId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
          nullable: true
        date:
          type: string
          format: date-time
        subtotal:
          type: number
          format: float
        tax:
          type: number
          format: float
        total:
          type: number
          format: float
        status:
          type: string
          enum: [pending, completed, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid]
        paymentId:
          type: string
          format: uuid
          nullable: true
...
    Error:
      type: object
      properties:
        message:
          type: string
paths:
  /auth/login:
    post:
      summary: Login a user (admin or cashier) and generate JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: Logout a user and invalidate JWT token
      tags:
        - Authentication
      security:
        - apiAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /sales:
    get:
      summary: Get a list of all sales transactions
      tags:
        - Sales
      security:
        - apiAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, cancelled]
          description: Filter sales by status
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter sales after this date (YYYY-MM-DD)
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter sales before this date (YYYY-MM-DD)
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter sales by customer ID
      responses:
        '200':
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sale'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new sale transaction
      tags:
        - Sales
      security:
        - apiAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                    required:
                      - productId
                      - quantity
                customerId:
                  type: string
                  format: uuid
              required:
                - items
      responses:
        '201':
          description: Sale created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sale:
                    $ref: '#/components/schemas/Sale'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product or customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sales/{id}:
    get:
      summary: Get details of a specific sale transaction
      tags:
        - Sales
      security:
        - apiAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the sale to retrieve
      responses:
        '200':
          description: Sale details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Sale'
                  - type: object
                    properties:
                      customer:
                        type: object
                        nullable: true
                        properties:
                          id:
                            type: string
                            format: uuid
                          name:
                            type: string
                          email:
                            type: string
                          phone:
                            type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Sale not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...